requirements.txt
Django
PyJWT
djangorestframework
djangorestframework-simplejwt
